Allow passing multiple args to Serializer.list
This allows using the convenience of ListSerializer with Serializers
that take multiple arguments on their initializer, with the assumption
that when rendering a list only the first argument changes, and other
arguments remain static.

An example is when a serializer must render information about certain
objects but in the context of the current user. The usual signature for
the serializer would be:

    class GameSerializer < Granola::Serializer
      def initialize(game, user)
        super(game)
        @user = user
      end

      def attributes
        {
          "name" => object.name,
          # ... other game specific things ...
          "wins" => object.wins(for: @user),
          "loses" => object.loses(for: @user),
        }
      end
    end

With this commit, if you wanted to just render a list of this
serializer, you can just do this:

    GameSerializer.list(games, user).to_json