Formalize Renderers as an object
Instead of having arbitrary hashes, keep objects with a defined
interface we can use, and hide the implementation / state.

Introduces Granola.renderer(type) to fetch a Renderer of that type.

This also deprecates Serializer#render (in favor of Renderer#render) and
Serializer#mime_type (in favor of Renderer#content_type). The methods
will be removed in v1.0.